requirements.txt
appwrite==11.0.0
python-dotenv
flask
